requirements.txt
# pip install -r requirements.txt

# 웹 애플리케이션 프레임워크
streamlit==1.46.1

# AI 및 머신러닝 라이브러리
sentence-transformers==5.0.0
transformers>=4.21.0
torch==2.7.1
langchain-upstage 
langchain-community 
sentence-transformers

# 데이터 처리
pandas==2.3.1
numpy==2.3.1

# 벡터 검색 엔진
faiss-cpu==1.11.0

# 환경 변수 관리
python-dotenv>=1.0.0

